feat: support submissions from app.tinytapeout.com (#2)
Previously, we had to fetch all the projects whenever we built Tiny Tapeout. This would make the builds inconsistent - if a user pushed a change around the same time we built TT-04, it was hard for them to tell whether their change made it to the Tiny Tapeout or not.

Now all the user modules are submitted through app.tinytapeout.com, which opens a new PR per submission, with all the macro files (gds/lef/spef/gate level net list). This PR changes the tinytapeout-04 actions to support this new format.

We also moved some of the functionality out of this repo into tt-support-tools, ditched the old Makefile and projects.yaml. Instead, we only use modules.yaml to specify the positions of some hard-coded macros. User macros are automatically detects and positioned during the GDS build process.

The GDS action now creates two new artifacts:

* efabless_submission - contains the files required for the Efabless chipIgnore submission.
* shuttle_index - a JSON file which lists of the projects in the shuttle, along with their mux addresses, repo, and commit hash.